backend/requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
python-jose
sqlalchemy
psycopg[binary]
pytest
pytest-asyncio
cryptography
httpx
alembic
